Hotfix issue #1520
NOT INTENDED TO BE USED GENERALLY!

Assumption: the tracked peers in InMemoryMessageIdProvider.trackers are
not removed on timeout.

Therefore use the 3.0 LeastRecentlyUsedCache.
Requires to add the ClockUtil and adapt InMemoryConnectionStore.
Update also SslContextUtil to use explicit TLSv1.2 in order to support
build with newer jdks.
Recreates the certificates for build.

Signed-off-by: Achim Kraus <achim.kraus@bosch.io>